requirements-dev.txt
# --- testes e2e ---
pytest==8.3.5
pytest-playwright==0.7.0
pytest-django==4.8.0
pytest-asyncio==0.23.6